start switch from msmtp-mta to postfix
Moving from msmtp-mta to postfix for null mailer because
msmtp-mta does not set date headers on messages.

This changes the packages installed at includes a basic
/etc/postfix/main.cf configured as a null mailer. It will
work if used with the mail_server role for another host in the
same playbook.

Still to do are allowing user to override relay host and use
SMTP-AUTH, and a handler so aliases are regenerated.