add auth support to postfix client, some issues
Added some variables that the templates will use if they exist to
allow the user to configure a relay host (and also username and
password for that host). I think the Postfix main.cf is done correctly,
but on my VMs that I use for testing I don't have the domains for the
host names set correctly so the ansible_fqdn variable is giving me
.localdomain. I need to fix my VMs before I can really test this
properly.