Mail delivery from client to server works
Solved some of the problems I was having with my VMs and got my
hostnames and FQDNs all set properly.

In postfix's main.cf I have the myorigin set to ansible_fqdn, which
makes it very clear which server a message is from. The message
arrives in the mailbox on the mail_server (as long as that inbox
exists).