Basic MoinMoin installation
Basic installation of a single MoinMoin wiki.

It would be nice to be able to support multiple wikis in a farm
config but I'm still thinking about the best way to do this. There's
a Python module /etc/moin/farmconfig.py that has a list of wikis
that are on the farm. It doesn't look like there's an easy way to
manage this except all in a single role. Maybe this role will
default to installing a single wiki, but allow more to be defined.