use default values for most variables
The templates have been updated to use the .get() method on
the dictionaries in moinmoin_wikis to provide reasonable defaults
and so the playbook does not fail if one of the dictionaries doesn't
have all of the keys defined.

Also fixed permissions on some config files that were unnecessarily
executable.