use custom SSL certs
The moinmoin role will now use the ssl_cert and ssl_key for items
in the moinmoin_wikis list. These should be the paths to the SSL
cert and key files to use.

If these are not supplied for a wiki, the playbook will still
configure https access using the snakeoil certificate that is
generated and self-signed by the Debian ssl-cert package.

If these are supplied, the role will configure Apache to
always redirect to https.