initial awstats support
I'm configuring awstats to generate static html reports instead
of serving them dynamically as a CGI script.

A new system user is created to run the update and build reports
scripts, but now it's got the reports in /var/cache/awstats
all owned by awstats. Either need to move them or change the
permissions on them so that Apache can serve them.

I might end up using modified copies of the scripts from the
Debian package that handle updating and building reports so
that the reports end up in a place that's more suitable than
/var/cache/awstats.